[mdf] 크게 각 PAGE별로 임시레이아웃과 기능 분배
