+++ LittleAchievement/account/views.py
@@ -8,17 +8,7 @@ from task.models import CommonTask,MyTask
 from django.contrib.auth import authenticate, login
 
 # Create your views here.
-def index(request):
-    context = dict()
-    active_user = request.user
-    context['all_common_task'] = CommonTask.objects.all()
 
-    if str(active_user) != "AnonymousUser":
-
-        context['all_my_task'] = MyTask.objects.filter(user = active_user, is_checked = False)
-        context['total_task'] = MyTask.objects.filter(user = active_user).count()
-        context['complete_task'] = MyTask.objects.filter(user = active_user, is_checked = True).count()
-    return render(request, 'index.html',context)
 
 def signup(request):
     if request.user.is_authenticated:
