+++ LittleAchievement/taskcreate/static/css/task.css
@@ -1,0 +1,8 @@
+.task_create_btn{
+    position: absolute;
+    right:1rem;
+    bottom:1rem;
+    border:2px solid green;
+    border-radius: 20px;
+
+}
