+++ LittleAchievement/taskcreate/templates/taskcreate.html
@@ -1,16 +1,16 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
+{%extends 'base.html'%}
+{%load static%}
+{% block extrahead %}
+<link rel="stylesheet" href="{% static 'css/task.css' %}">
+
+{% endblock %}
+
+{%block content%}
+
     <h1> Task Creates</h1>
     <form method="POST"> 
         {%csrf_token%}
         {{form.as_p}}
         <button type="submit"> 제출 </button>
     </form>
-</body>
-</html>
+{%endblock%}
